requirements.txt
streamlit==1.41.1
python-dotenv==1.0.1

langchain==0.2.16
langchain-community==0.2.16
langchain-text-splitters==0.2.4
langchain-groq==0.1.9

chromadb==0.5.5
sentence-transformers==3.0.1

pypdf==4.3.1
docx2txt==0.8
